requirements.txt
python-telegram-bot
requests
google-auth-httplib2
google-api-python-client
google-auth
google-auth-oauthlib